[REFACTOR] #43 - 리스트화
