+++ app/src/main/java/org/woowatechcamp/githubapplication/data/issue/model/IssueResponse.kt
@@ -1,3 +1,50 @@
 package org.woowatechcamp.githubapplication.data.issue.model
 
-class IssueResponse : ArrayList<IssueResponseItem>()
+import org.woowatechcamp.githubapplication.presentation.home.issue.model.IssueModel
+import org.woowatechcamp.githubapplication.presentation.home.issue.model.IssueState
+import org.woowatechcamp.githubapplication.util.ext.getDate
+import org.woowatechcamp.githubapplication.util.ext.getIndexString
+import org.woowatechcamp.githubapplication.util.ext.getTimeDiff
+
+data class IssueResponse(
+    val active_lock_reason: Any,
+    val assignee: Assignee,
+    val assignees: List<Assignee>,
+    val author_association: String,
+    val body: String,
+    val closed_at: Any,
+    val comments: Int,
+    val comments_url: String,
+    val created_at: String,
+    val draft: Boolean,
+    val events_url: String,
+    val html_url: String,
+    val id: Int,
+    val labels: List<Label>,
+    val labels_url: String,
+    val locked: Boolean,
+    val milestone: Any,
+    val node_id: String,
+    val number: Int,
+    val performed_via_github_app: Any,
+    val pull_request: PullRequest,
+    val reactions: Reactions,
+    val repository: Repository,
+    val repository_url: String,
+    val state: String,
+    val state_reason: Any,
+    val timeline_url: String,
+    val title: String,
+    val updated_at: String,
+    val url: String,
+    val user: User
+) {
+    fun getModel() = IssueModel(
+        state = IssueState.getIssueState(state),
+        name = repository.name,
+        fullName = repository.full_name,
+        number = number.getIndexString(),
+        title = title,
+        timeDiff = updated_at.getDate().getTimeDiff()
+    )
+}
