+++ app/src/main/java/org/woowatechcamp/githubapplication/data/noti/model/NotiResponse.kt
@@ -1,5 +1,16 @@
-package org.woowatechcamp.githubapplication.data.notifications.model
+package org.woowatechcamp.githubapplication.data.noti.model
 
-import org.woowatechcamp.githubapplication.data.noti.model.NotiResponseItem
+import org.woowatechcamp.githubapplication.data.notifications.model.Repository
+import org.woowatechcamp.githubapplication.data.notifications.model.Subject
 
-class NotiResponse : ArrayList<NotiResponseItem>()
+data class NotiResponse(
+    val id: String,
+    val last_read_at: String,
+    val reason: String,
+    val repository: Repository,
+    val subject: Subject,
+    val subscription_url: String,
+    val unread: Boolean,
+    val updated_at: String,
+    val url: String
+)
